% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_thaliana_bmData}
\alias{A_thaliana_bmData}
\title{Simulated arabidopsis thaliana bmData object.}
\format{
An object of class \code{bmData} with 40001 rows and 2 columns.
}
\usage{
A_thaliana_bmData
}
\description{
A simulated data set contains simulated IPO information.
}
\keyword{datasets}
