% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_thaliana_BSobj}
\alias{A_thaliana_BSobj}
\title{Arabidopsis thaliana BSseq object created by DSS}
\format{
An object of class \code{BSseq} with 40001 rows and 2 columns.
}
\usage{
A_thaliana_BSobj
}
\description{
A data set contains methylation information
}
\keyword{datasets}
