% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_thaliana_dmR}
\alias{A_thaliana_dmR}
\title{Different methylation region created by DSS}
\format{
A data frame with 1 row and 9 variables
\describe{
\item{chr}{chromosome, the chromosome information of dmR}
\item{start}{the start site of dmR}
\item{end}{the end site of dmR}
\item{length}{the length of dmR}
\item{nCG}{Number of CpG sites contained in the DMR}
\item{meanMethy1}{Average methylation levels in two conditions}
\item{meanMethy2}{Average methylation levels in two conditions}
\item{diff.Methy}{The difference in the methylation levels between two conditions}
\item{areaStat}{	The sum of the test statistics of all CpG sites within the DMR}
}
}
\usage{
A_thaliana_dmR
}
\description{
A data set contains information of different methylation region
see data-row/A_thaliana_data_procession.R
}
\keyword{datasets}
