% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bmData-class}
\alias{bmData-class}
\title{bmData Class}
\description{
This class added extra data to BSseq class. Change the assays by storing
M/Cov to any value1/2
}
\seealso{
bmData class inherits RangedSummarizedExperiment class,
other slots see \linkS4class{RangedSummarizedExperiment}
}
\keyword{classes}
