% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMplot-class.R
\docType{class}
\name{BMplot-class}
\alias{BMplot-class}
\alias{.BMplot}
\alias{BMplot}
\title{Class BMplot
This class is a subclass from BSseq class, storing the information of any
value of interest. Acting as a reference role for visualizing base modification}
\description{
Class BMplot
This class is a subclass from BSseq class, storing the information of any
value of interest. Acting as a reference role for visualizing base modification
}
\section{Slots}{

\describe{
\item{\code{valuenames}}{the name of each value stored in BMplot object}
}}

\keyword{classes}
