% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMplot-package.R
\docType{package}
\name{BMplot-package}
\alias{BMplot-package}
\alias{BMplot}
\title{Base Modification visualization}
\description{
This package is designed for visualizing base modification
}
\details{
\tabular{ll}{
Package: \tab BMplot\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 23-01-2022\cr
License: \tab Artistic-2.0\cr }
}
\author{
Ming Li

Maintainer: Ming Li \email{limiang929@gmail.com}
}
\keyword{package}
