% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseModificationDf.R
\name{create_regex_patterns_negative}
\alias{create_regex_patterns_negative}
\title{create regex patterns in negative strand}
\usage{
create_regex_patterns_negative(motif)
}
\arguments{
\item{motif}{the motif(e.g C:CG/CH, A:GAGG/AGG) of the base modification}
}
\description{
create regex patterns in negative strand
}
