% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_regex_patterns_positive}
\alias{create_regex_patterns_positive}
\title{create regex patterns in positive strand}
\usage{
create_regex_patterns_positive(motif)
}
\arguments{
\item{motif}{the motif(e.g C:CG/CH, A:GAGG/AGG) of the base modification}
}
\description{
create regex patterns in positive strand
}
