% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseModificationDf.R
\name{getBaseModificationDf.bmData}
\alias{getBaseModificationDf.bmData}
\title{get the information of base modification}
\usage{
getBaseModificationDf.bmData(
  region,
  input,
  BSgenome,
  base = NULL,
  motif = NULL,
  position_bias = NULL
)
}
\arguments{
\item{region}{base modification region in the form of dataframe, having columns of "chr","start" and "end"}

\item{input}{the input data stored in bmData objects}

\item{BSgenome}{genome reference}

\item{base}{one of A/T/G/C/U}

\item{motif}{the motif(e.g C:CG/CH, A:GAGG/AGG) of the base modification}

\item{position_bias}{1-base bias. e.g position_bias = 1("C" in "CHH"), position_bias = 2("A" in "GAGG")}
}
\value{
dataframe
}
\description{
This function retrieve the information of each base, requiring bmData object as input.
Then organized it to dataframe.
}
