% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseModificationDf.R
\name{getBaseModificationDf.BSseq}
\alias{getBaseModificationDf.BSseq}
\title{get the information of base modification}
\usage{
getBaseModificationDf.BSseq(
  region,
  input,
  BSgenome,
  strand = TRUE,
  cover_depth = TRUE,
  base = NULL,
  motif = NULL,
  position_bias = NULL
)
}
\arguments{
\item{region}{base modification region in the form of dataframe, having columns of "chr","start" and "end"}

\item{input}{the input data stored in BSseq objects or BSseqExtra objects}

\item{BSgenome}{genome reference}

\item{strand}{distinguish strand information or not}

\item{cover_depth}{take the depth of cover into account or not}

\item{base}{one of A/T/G/C/U}

\item{motif}{the motif(e.g C:CG/CH, A:GAGG/AGG) of the base modification}

\item{position_bias}{1-base bias. e.g position_bias = 1("C" in "CHH"), position_bias = 2("A" in "GAGG")}
}
\value{
dataframe
}
\description{
This function retrieve the information of each base, requiring BSseq object as input.
Then organized it to dataframe.
}
