% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Human_BSobj}
\alias{Human_BSobj}
\title{Human BSseq object created by DSS}
\format{
An object of class \code{BSseq} with 30002 rows and 4 columns.
}
\usage{
Human_BSobj
}
\description{
A data set contains methylation information
}
\keyword{datasets}
