% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebmData.R
\name{makebmDataFromData.internal}
\alias{makebmDataFromData.internal}
\title{makebmDataFromData.internal}
\usage{
makebmDataFromData.internal(data, sampleNames = NULL)
}
\arguments{
\item{data}{lists object}

\item{sampleNames}{the name of each samples}
}
\value{
dmData object
}
\description{
make dmData object from data
}
\details{
This internal function was inspired by DSS::makeBSseqData.

The objects in \code{data} must have specific forms. Colunms should be
features, which should be organized in the order of "chr", "pos", "value1",
"value2(optional)". chr stands for chromosome. pos stands for position on
chromosome, also known as coordinates. value1/2 stands for the value on each base.
The colnames can be any character but must be in the order. Rows stands for each
observation.
}
