% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseModificationProf.R
\name{plotBaseModificationProf}
\alias{plotBaseModificationProf}
\title{plotBaseModificationProf}
\usage{
plotBaseModificationProf(
  df,
  motif_color = NULL,
  title = NULL,
  xlim = NULL,
  GeneModel = NULL,
  highlight_lim = NULL,
  highlight_color = NULL,
  highlight_alpha = 0.013,
  xlab = "Genomic Region(5'->3')",
  ylab = NULL,
  second_ylab = NULL,
  switch_y_value = FALSE,
  legend_lab_motif = NULL,
  legend_lab_value2 = NULL,
  switch_facet_label = TRUE,
  strip_placement = "outside",
  angle_of_facet_label = 0,
  alpha = 0.6,
  y_ticks_length = 0.25,
  x_ticks_length = 0.25,
  strip_border = FALSE,
  facet_label_text_size = 12,
  axis_title_text_size = 17,
  title_text_size = 20,
  right_y_axis_text_size = 13,
  left_y_axis_text_size = 13,
  x_axis_text_size = 13,
  depth_heatmap = TRUE,
  nrow = NULL,
  ncol = NULL,
  panel_spacing = 1,
  legend_box_spacing = 3,
  legend_position = "right"
)
}
\arguments{
\item{df}{the base modification dataframe}

\item{motif_color}{the color for different motifs(CHH,CHG,CG)}

\item{title}{the title of the plot, can also be a list of title}

\item{xlim}{the specified interval of region, must be the sub-interval of the dmR. list for list df}

\item{GeneModel}{annotation data, can be TxDb/GRanges/GRangesList etc. Details see \code{\link[ggbio:autoplot-method]{ggbio::autoplot()}}}

\item{highlight_lim}{the high light region. list for list df}

\item{highlight_color}{the color of high light rectangular box}

\item{highlight_alpha}{the alpha of highlight box}

\item{xlab}{the x label, can also be a list of x label}

\item{ylab}{the y label, can also be a list of y label}

\item{second_ylab}{the ylab for second y-axis}

\item{switch_y_value}{switch the value from left y-axis to right y-axis}

\item{legend_lab_motif}{the label of legend for motif}

\item{legend_lab_value2}{the label of legend for the second value(ylab is the label for the first value)}

\item{switch_facet_label}{switch the facet label from right to left or not}

\item{strip_placement}{strip.placement}

\item{angle_of_facet_label}{the angle of facet label, e.g. 0 is horizontal}

\item{alpha}{transparency for the depth information line}

\item{y_ticks_length}{the length of y-axis ticks}

\item{x_ticks_length}{the length of x-axis ticks}

\item{strip_border}{add border to the facet label or not}

\item{facet_label_text_size}{the size of facet label text}

\item{axis_title_text_size}{the size of axis title text}

\item{title_text_size}{the size of the title text}

\item{right_y_axis_text_size}{the size of the left y axis text,this work when depth information is taken into account}

\item{left_y_axis_text_size}{the size of the left y axis text}

\item{x_axis_text_size}{the size of x axis text}

\item{depth_heatmap}{draw the heatmap of depth information or not}

\item{nrow}{the nrow of plotting a list of dmR}

\item{ncol}{the ncol of plotting a list of dmR}

\item{panel_spacing}{the distance between panels}

\item{legend_box_spacing}{the distance between legend and plotting area,"cm"}

\item{legend_position}{the position of legend}
}
\value{
ggplot object
}
\description{
plot base modification profile
}
