% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseModificationDf.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeBSseqData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DSS}{\code{\link[DSS]{makeBSseqData}}}
}}

