% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SamepleBedFiles}
\alias{SamepleBedFiles}
\title{Get the sample bed files}
\usage{
SamepleBedFiles()
}
\description{
This function overwrite the ChIPseeker::getSampleFiles().
}
\seealso{
\code{\link[ChIPseeker]{getSampleFiles}}
}
