% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SamepleTxtFiles}
\alias{SamepleTxtFiles}
\title{Get the sample txt files}
\usage{
SamepleTxtFiles()
}
\description{
This function overwrite the ChIPseeker::getSampleFiles().
}
\seealso{
\code{\link[ChIPseeker]{getSampleFiles}}
}
