% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_BSobj}
\alias{simulated_BSobj}
\title{Simulated arabidopsis thaliana BSseq object created by DSS}
\format{
An object of class \code{BSseq} with 523 rows and 2 columns.
}
\usage{
simulated_BSobj
}
\description{
A simulated data set contains methylation information.
These data were created by finding the adenine positions in a specific region,
which was supposed to be the dmR, and assign the adenine sites with random values.
At last created BSseq object by DSS.
}
\keyword{datasets}
