% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_dmR}
\alias{simulated_dmR}
\title{Different methylation region created by simulation}
\format{
A data frame with 1 row and 3 variables
\describe{
\item{chr}{chromosome, the chromosome information of dmR}
\item{start}{the start site of dmR}
\item{end}{the end site of dmR}
}
}
\usage{
simulated_dmR
}
\description{
A data set contains information of different methylation region
see data-row/simulated_A_thaliana_data_procession.R
}
\keyword{datasets}
