% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_IPO}
\alias{simulated_IPO}
\title{Simulated IPO data of A_thaliana}
\format{
An object of class \code{list} of length 2.
}
\usage{
simulated_IPO
}
\description{
A data set contains simulated IPO information. This data set is simulated
from GSE62206. We use two samples from GSE62206 and replace the previous value
in GSE62206 with random IPO and IPO ratio.
see data-row/A_thaliana_IPO_data_procession.R
}
\keyword{datasets}
