% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseModificationDf.R
\name{create_regex_patterns_positive}
\alias{create_regex_patterns_positive}
\title{create regex patterns in positive strand}
\usage{
create_regex_patterns_positive(motif)
}
\description{
create regex patterns in positive strand
}
